+++ test/change_health/response/claim/report/report_835_data_test.rb
@@ -45,6 +45,18 @@ class Report835DataTest < Minitest::Test
         assert_equal report_name, actual_payment.report_name
         assert_equal '810.8', actual_payment.total_actual_provider_payment_amount
       end
+
+      it 'payment provider adjustments' do
+        assert_equal 1, actual_payment.provider_adjustments.size
+
+        provider_adjustment = actual_payment.provider_adjustments[0]
+
+        assert_equal '1124058920', provider_adjustment.provider_identifier
+        assert_equal Date.new(2022, 12, 31), provider_adjustment.fiscal_period_date
+        adjustments = [{ amount: '52436.08', identifier: '20211124 XP732039', reason_code: 'WO' },
+                       { amount: '-49082.64', identifier: '9802717391', reason_code: 'FB' }]
+        assert_equal adjustments, provider_adjustment.adjustments
+      end
     end
 
     describe 'claims' do
@@ -87,6 +99,7 @@ class Report835DataTest < Minitest::Test
 
         expected_claim = ChangeHealth::Response::Claim::Report835Claim.new(
           claim_payment_remark_codes: ['N520'],
+          claim_status_code: '1',
           patient_control_number: '7722337',
           patient_first_name: 'SANDY',
           patient_last_name: 'DOE',
@@ -140,6 +153,7 @@ class Report835DataTest < Minitest::Test
       describe 'claim field oddities' do
         let(:odd_claim1) { report_data.payments[1].claims[0] }
         let(:odd_claim2) { report_data.payments[2].claims[0] }
+        let(:odd_claim_service_date) { report_data.payments[2].claims[2] }
         it 'member id' do
           assert_equal 'SJD11122', odd_claim1.patient_member_id
         end
@@ -150,6 +164,11 @@ class Report835DataTest < Minitest::Test
 
         it 'provider npi from payee npi' do
           assert_equal '9999947036', odd_claim2.service_provider_npi
+        end
+
+        it 'service line serviceDate missing' do
+          assert_equal Date.new(2019, 3, 23), odd_claim_service_date.service_date_begin
+          assert_equal Date.new(2019, 3, 24), odd_claim_service_date.service_date_end
         end
       end
     end