+++ CHANGELOG.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [4.6.0] - 2022-09-01
+
+### Added
+
+* Report835Data - extract claimStatusCode for Report835Claim
+
+# [4.5.0] - 2022-08-29
+
+### Added
+
+* Report835Data - pull out provider adjustments
+
+# [4.4.0] - 2022-08-29
+
+### Added
+
+* Claim Submission - add payToAddress field
+
+# [4.3.0] - 2022-08-08
+
+### Added
+
+* Report835Data - add another way to get service_date_begin & service_date_end
+
+# [4.2.5] - 2022-07-29
+
+### Fixed
+* Report835Data can now handle when a service line doesn't have a service date
+
 # [4.2.4] - 2022-07-28
 
 * Removed all RARC Codes (HealthCareCheckRemarkCodes) from JSON
@@ -401,6 +430,11 @@ Added the ability to hit professional claim submission API. For more details, se
 * Authentication
 * Configuration
 
+[4.6.0]: https://github.com/WeInfuse/change_health/compare/v4.5.0...v4.6.0
+[4.5.0]: https://github.com/WeInfuse/change_health/compare/v4.4.0...v4.5.0
+[4.4.0]: https://github.com/WeInfuse/change_health/compare/v4.3.0...v4.4.0
+[4.3.0]: https://github.com/WeInfuse/change_health/compare/v4.2.5...v4.3.0
+[4.2.5]: https://github.com/WeInfuse/change_health/compare/v4.2.4...v4.2.5
 [4.2.4]: https://github.com/WeInfuse/change_health/compare/v4.2.3...v4.2.4
 [4.2.3]: https://github.com/WeInfuse/change_health/compare/v4.2.2...v4.2.3
 [4.2.2]: https://github.com/WeInfuse/change_health/compare/v4.2.1...v4.2.2